+++ src/assets/data/projects.jsx
@@ -1,0 +1,57 @@
+export const projectsData = [
+    {
+      id: "lucifer",
+      name: "LUCIFER'S COLOSSEUM",
+      site: "https://checkmatechris1.itch.io/lucifers-colosseum",
+      type: "GAME",
+      picture: "/lc.png",
+      imageShape: "wide",
+      tags: ["GDScript", "Blender", "Audacity", "Godot"],
+      abstract:
+        "Lucifer's Colosseum is a round-based, rogue-like, and arena-based FPS game built in Godot. It was a submission to the UCI Winter 2026 Video Game Design Club Games.\n\nI participated in a 11 person team of programmers, 3D/2D artists, audio engineers, UI/UX designers, and game designers using the Kanban system.",
+    },
+    {
+      id: "pcos",
+      name: "PCOSPositive",
+      site: "https://example.com",
+      type: "web",
+      picture: "/pcos.png",
+      imageShape: "square",
+      tags: ["Swift", "CoreData", "Figma"],
+      abstract:
+        "This is an example of a vertical image layout where the badges run down the middle column and the text sits on the far right. Notice how the grid completely morphs to accommodate the mobile aspect ratio!",
+    },
+    {
+      id: "crosswalk",
+      name: "CrossWalk",
+      site: "https://example.com",
+      type: "Mobile",
+      picture: "/cw.png",
+      imageShape: "vertical",
+      tags: ["Figma"],
+      abstract:
+        "This is an example of a vertical image layout where the badges run down the middle column and the text sits on the far right. Notice how the grid completely morphs to accommodate the mobile aspect ratio!",
+    },
+    {
+      id: "stamps",
+      name: "fu-stamps",
+      site: "https://example.com",
+      type: "Web",
+      picture: "/stamps.png",
+      imageShape: "wide",
+      tags: ["Swift", "CoreData", "Figma"],
+      abstract:
+        "This is an example of a vertical image layout where the badges run down the middle column and the text sits on the far right. Notice how the grid completely morphs to accommodate the mobile aspect ratio!",
+    },
+    {
+      id: "portfolio",
+      name: "Portfolio",
+      site: "",
+      type: "?",
+      picture: "/page.png",
+      imageShape: "wide",
+      tags: ["Tailwind", "React", "Figma"],
+      abstract:
+        "This is an example of a vertical image layout where the badges run down the middle column and the text sits on the far right. Notice how the grid completely morphs to accommodate the mobile aspect ratio!",
+    },
+  ];
