--- src/assets/organizations.jsx
@@ -1,57 +1,0 @@
-export const organizationsData = [
-    {
-      id: "ctc",
-      name: "COMMIT THE CHANGE",
-      site: "https://ctc-uci.com",
-      type: "Present",
-      picture: "/ctc.png",
-      imageShape: "wide",
-      tags: ["React", "ExpressJS", "AWS S3", "PostgreSQL", "Firebase"],
-      abstract:
-        "Commit the Change (CTC) is a collaborative development team that specializes in building software for nonprofit organizations.\n\nI am currently working on a management dashboard for Global Creation Foundation, a nonprofit focused on arts education in underserved communities that has reached 300+ students across 5 continents.\n\nI've integrated media uploads through AWS S3 buckets, created reusable ExpressJS routes, and much more through working with a 14 person Agile Scrum team. ",
-    },
-    {
-      id: "fusion",
-      name: "FUSION ICS PROJECT",
-      site: "https://github.com/JasonNguyen067/PinoyPort",
-      type: "Present",
-      picture: "/fusion.png",
-      imageShape: "square",
-      tags: ["React", "Supabase"],
-      abstract:
-        "Commit the Change (CTC) is a collaborative development team that specializes in building software for nonprofit organizations.\n\nI am currently working on a management dashboard for Global Creation Foundation, a nonprofit focused on arts education in underserved communities that has reached 300+ students across 5 continents.\n\nI've integrated media uploads through AWS S3 buckets, created reusable ExpressJS routes, and much more through working with a 14 person Agile Scrum team. ",
-    },
-    {
-      id: "zotbins",
-      name: "ZotBins",
-      site: "https://github.com/zotbins",
-      type: "Present",
-      picture: "/zotbins.png",
-      imageShape: "vertical",
-      tags: ["Supabase"],
-      abstract:
-        "Commit the Change (CTC) is a collaborative development team that specializes in building software for nonprofit organizations.\n\nI am currently working on a management dashboard for Global Creation Foundation, a nonprofit focused on arts education in underserved communities that has reached 300+ students across 5 continents.\n\nI've integrated media uploads through AWS S3 buckets, created reusable ExpressJS routes, and much more through working with a 14 person Agile Scrum team. ",
-    },
-    {
-      id: "game",
-      name: "Video Game Design",
-      site: "https://vgdc-uci.framer.website/",
-      type: "2026",
-      picture: "/vgdc.png",
-      imageShape: "square",
-      tags: [],
-      abstract:
-        "Commit the Change (CTC) is a collaborative development team that specializes in building software for nonprofit organizations.\n\nI am currently working on a management dashboard for Global Creation Foundation, a nonprofit focused on arts education in underserved communities that has reached 300+ students across 5 continents.\n\nI've integrated media uploads through AWS S3 buckets, created reusable ExpressJS routes, and much more through working with a 14 person Agile Scrum team. ",
-    },
-    {
-      id: "design",
-      name: "Design @ UCI",
-      site: "https://designatuci.com/",
-      type: "2025",
-      picture: "/dau.png",
-      imageShape: "square",
-      tags: [],
-      abstract:
-        "Commit the Change (CTC) is a collaborative development team that specializes in building software for nonprofit organizations.\n\nI am currently working on a management dashboard for Global Creation Foundation, a nonprofit focused on arts education in underserved communities that has reached 300+ students across 5 continents.\n\nI've integrated media uploads through AWS S3 buckets, created reusable ExpressJS routes, and much more through working with a 14 person Agile Scrum team. ",
-    },
-  ];
