+++ src/index.css
@@ -2,56 +2,56 @@
 
 @font-face {
   font-family: 'PPFraktionMono';
-  src: url('./assets/PPFraktionMono-Regular-BF675904a6a1564.otf') format('opentype');
+  src: url('./assets/fonts/PPFraktionMono-Regular-BF675904a6a1564.otf') format('opentype');
   font-weight: 400;
   font-style: normal;
 }
 
 @font-face {
   font-family: 'PPFraktionMono';
-  src: url('./assets/PPFraktionMono-RegularItalic-BF675904a6859b6.otf') format('opentype');
+  src: url('./assets/fonts/PPFraktionMono-RegularItalic-BF675904a6859b6.otf') format('opentype');
   font-weight: 400;
   font-style: italic;
 }
 
 @font-face {
   font-family: 'PPFraktionMono';
-  src: url('./assets/PPFraktionMono-Bold-BF675904a6c4c6e.otf') format('opentype');
+  src: url('./assets/fonts/PPFraktionMono-Bold-BF675904a6c4c6e.otf') format('opentype');
   font-weight: 700;
   font-style: normal;
 }
 
 @font-face {
   font-family: 'PPFraktionMono';
-  src: url('./assets/PPFraktionMono-BoldItalic-BF675904a6f3fc1.ttf') format('truetype');
+  src: url('./assets/fonts/PPFraktionMono-BoldItalic-BF675904a6f3fc1.ttf') format('truetype');
   font-weight: 700;
   font-style: italic;
 }
 
 @font-face {
   font-family: 'PPFraktionSans';
-  src: url('./assets/PPFraktionSans-Light-BF675904a63027f.otf') format('opentype');
+  src: url('./assets/fonts/PPFraktionSans-Light-BF675904a63027f.otf') format('opentype');
   font-weight: 300;
   font-style: normal;
 }
 
 @font-face {
   font-family: 'PPFraktionSans';
-  src: url('./assets/PPFraktionSans-LightItalic-BF675904a6ee54f.otf') format('opentype');
+  src: url('./assets/fonts/PPFraktionSans-LightItalic-BF675904a6ee54f.otf') format('opentype');
   font-weight: 300;
   font-style: italic;
 }
 
 @font-face {
   font-family: 'PPFraktionSans';
-  src: url('./assets/PPFraktionSans-Bold-BF675904a667d89.otf') format('opentype');
+  src: url('./assets/fonts/PPFraktionSans-Bold-BF675904a667d89.otf') format('opentype');
   font-weight: 700;
   font-style: normal;
 }
 
 @font-face {
   font-family: 'PPFraktionSans';
-  src: url('./assets/PPFraktionSans-BoldItalic-BF675904a70061a.otf') format('opentype');
+  src: url('./assets/fonts/PPFraktionSans-BoldItalic-BF675904a70061a.otf') format('opentype');
   font-weight: 700;
   font-style: italic;
 }
